TXN-1102: Handle signing transactions for multiple connected accounts (#14)
Before sending transactions to the wallet to be signed, we check if the transaction sender is the same as the active address. This failed to take into account that some wallets supported connecting multiple accounts, and limited the user to only signing transactions for the active address. This fixes that issue by checking against all connected accounts, instead of just the active address.