fix(walletconnect2): catch "Modal closed" error in connect method
Closing the Web3Modal was throwing an uncaught error instead of rejecting the promise returned by
`connect`, leading to a runtime error. This refactors the connection logic using then-catch chaining
(using async/await would lead to `connect` returning  `Promise<Promise<Wallet>>`) so the error is
caught and the promise is rejected.