feat(init): static imports are required
The cleanest, simplest solution to the build/compile errors encountered if any peer dependencies are
not installed is to require static imports. This is already a requirement for Remix apps, and it
will ensure that only the necessary wallet client/SDK libraries need to be installed by consuming
apps. This also removes the need for a "default configuration" and deciding which providers should
or shouldn't belong in it. Every wallet provider will need to be passed to the `providers` property
in useInitializeProviders, which is already how most apps do it.

BREAKING CHANGE: The `useInitializeProviders` must receive an object with the `providers` property
defined. Provider objects passed to `providers` must include a `clientStatic` property set to the
provider's client/SDK library, if it exists. For example, you cannot simply pass `PROVIDER_ID.PERA`,
it must be an object with both `id` and `clientStatic` set.