TXN-1412: Static imports are required (#90)
* feat(init): static imports are required

The cleanest, simplest solution to the build/compile errors encountered if any peer dependencies are
not installed is to require static imports. This is already a requirement for Remix apps, and it
will ensure that only the necessary wallet client/SDK libraries need to be installed by consuming
apps. This also removes the need for a "default configuration" and deciding which providers should
or shouldn't belong in it. Every wallet provider will need to be passed to the `providers` property
in useInitializeProviders, which is already how most apps do it.

BREAKING CHANGE: The `useInitializeProviders` must receive an object with the `providers` property
defined. Provider objects passed to `providers` must include a `clientStatic` property set to the
provider's client/SDK library, if it exists. For example, you cannot simply pass `PROVIDER_ID.PERA`,
it must be an object with both `id` and `clientStatic` set.

* test(init): fix failing tests

* fix(init): update NodeConfig type

The `nodePort` property can be a string or a number, and the optional `nodeHeaders` property is
added

* docs: update README

Updates code examples, provider configuration section, and migration guide to describe required
static imports. Various other changes as well.

* chore(example): update Storybook example to useInitializeProviders

* test(example): fix failing tests

* chore: rename ProvidersArray type

* chore(init): remove initializeProviders main export

The `useInitializeProviders` hook replaces `initializeProviders` as a helper function for
constructing the providers map that is passed to the context provider in consuming apps.

BREAKING CHANGE: Apps using `initializeProviders` will be forced to switch to
`useInitializeProviders`