refactor(types): single InitParams type that uses ProviderConfigMapping
This actually started as an attempt to resolve "circular dependencies" warnings that started
appearing after upgrading Rollup. It resulted in a refactored `InitParams` which extends
`CommonInitParams` with the `clientOptions` and `clientStatic` for each provider in
`ProviderConfigMapping`. It also resolves all but two of the warnings (will need to revisit) and
organizes the imports for each of the provider classes.