TXN-1414: Refactor InitParams type, tidy up provider class imports (#92)
* chore(deps): upgrade zustand to ^4.3.8

This also updates how `create` is imported in the stores. The default import has been deprecated and
shows a warning if used in local dev.

* refactor(types): single InitParams type that uses ProviderConfigMapping

This actually started as an attempt to resolve "circular dependencies" warnings that started
appearing after upgrading Rollup. It resulted in a refactored `InitParams` which extends
`CommonInitParams` with the `clientOptions` and `clientStatic` for each provider in
`ProviderConfigMapping`. It also resolves all but two of the warnings (will need to revisit) and
organizes the imports for each of the provider classes.