fix(walletconnect): get default metadata from window
The @walletconnect/utils package exports a function called `getWindowMetadata` which inspects the HTML Document for meta tags, favicon, etc, to populate `metadata` properties if values are not set in the configuration.

This adds a private method to the WalletConnect provider adapted from the implementation from the package to avoid requiring it as a dependency (or peer dependency).

When installing the package as a dependency, tree-shaking does not seem to have an effect. Despite only using this one function, the entire library gets included in the bundle, more than doubling use-wallet's bundle size.