chore: revert Pera provider to Pera Connect v1
While the testing for the Pera Connect v2 beta so far has been very promising, due to an unresolved issue this reverts the Pera provider back to Pera Connect v1.

A dependency of `@perawallet/connect-beta` – `@walletconnect/sign-client` is not included in Vite production builds. Vite's dev mode uses esbuild and everything works fine, but for some reason the Rollup production builds omit the Sign client and an error is thrown when you try to connect.

Instead of removing the new provider, it will remain as an undocumented option, `PERA2`, so it can continue to be tested on an opt-in basis.