fix(core): fixes isTransactionArray utility function
In a controlled test environment using `instanceof` to verify a transaction is of type `algosdk.Transaction` works, but in a production setting a more "low tech" type guard is necessary.

Now the items (either a normal transaction object or encoded transaction) are checked for a `genesisID` property, which confirms it is an unencoded object.